requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
aiofiles>=23.2.1
pyyaml>=6.0.1
rich>=13.7.0
click>=8.1.7
jinja2>=3.1.4
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
lxml>=5.2.0
dnspython>=2.6.1
python-whois>=0.9.4
cryptography>=42.0.0
web3>=7.0.0
requests>=2.32.0
